Rename: seniors.domain.common -> seniors.common.entity 변경 #95
