+++ src/main/java/com/seniors/domain/chat/entity/ChatMessage.java
@@ -1,6 +1,6 @@
 package com.seniors.domain.chat.entity;
 
-import com.seniors.domain.common.BaseEntity;
+import com.seniors.common.entity.BaseEntity;
 import com.seniors.domain.users.entity.Users;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
