+++ src/main/java/com/seniors/domain/notification/entity/Notification.java
@@ -1,6 +1,6 @@
 package com.seniors.domain.notification.entity;
 
-import com.seniors.domain.common.BaseTimeEntity;
+import com.seniors.common.entity.BaseTimeEntity;
 import com.seniors.domain.users.entity.Users;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
