+++ src/main/java/com/seniors/domain/post/entity/PostLike.java
@@ -1,6 +1,6 @@
 package com.seniors.domain.post.entity;
 
-import com.seniors.domain.common.BaseTimeEntity;
+import com.seniors.common.entity.BaseTimeEntity;
 import com.seniors.domain.users.entity.Users;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
