+++ src/main/java/com/seniors/domain/resume/entity/ResumeView.java
@@ -1,7 +1,6 @@
 package com.seniors.domain.resume.entity;
 
-import com.seniors.domain.common.BaseTimeEntity;
-import com.seniors.domain.resume.dto.ResumeDto;
+import com.seniors.common.entity.BaseTimeEntity;
 import com.seniors.domain.users.entity.Users;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
